ci(bench): activate benchmark gate evidence workflow
Adds benchmark gate workflow activation, release benchmark wrapper, native preflight/live-smoke scripts, gated baselines, Starlark compute benchmark support, and source-visible H0-H9 checkpoint changes. Excludes local docs-dev evidence and generated target artifacts per H9 policy.

+++ crates/vil_server_core/src/otel.rs
@@ -29,7 +29,7 @@ impl TraceId {
         let hi = std::time::SystemTime::now()
             .duration_since(SystemTime::UNIX_EPOCH)
             .unwrap_or_default()
-            .as_nanos() as u128;
+            .as_nanos();
         Self(hi ^ (SPAN_COUNTER.fetch_add(1, Ordering::Relaxed) as u128))
     }
